+++ src/main/java/com/umc/linkyou/web/controller/LinkuController.java
@@ -7,6 +7,7 @@ import com.umc.linkyou.apiPayload.exception.handler.UserHandler;
 import com.umc.linkyou.config.security.jwt.CustomUserDetails;
 import com.umc.linkyou.converter.LinkuConverter;
 import com.umc.linkyou.service.Linku.LinkuCreateService;
+import com.umc.linkyou.service.Linku.LinkuRecommendService;
 import com.umc.linkyou.service.Linku.LinkuSearchService;
 import com.umc.linkyou.service.Linku.LinkuService;
 import com.umc.linkyou.utils.UsersUtils;
@@ -32,6 +33,7 @@ public class LinkuController {
     private final LinkuService linkuService;
     private final LinkuCreateService linkuCreateService;
     private final LinkuSearchService linkuSearchService;
+    private final LinkuRecommendService linkuRecommendService;
     private final UsersUtils usersUtils;
 
     @PostMapping(value = "", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
@@ -109,7 +111,7 @@ public class LinkuController {
             @RequestParam(defaultValue = "5") int size
     ) {
         Long userId = usersUtils.getAuthenticatedUserId(userDetails);
-        return linkuService.recommendLinku(userId, situationId, emotionId, page, size);
+        return linkuRecommendService.recommendLinku(userId, situationId, emotionId, page, size);
     }//linku 추천 내부로
 
     // 빠른 검색 (사용자가 저장한 링크 전체 대상)
